+++ src/app/common/application/language.ts
@@ -1,3 +1,3 @@
 export class Language {
-    constructor(public code: string, public englishName: string, public localizedName: string) {}
+    constructor(public code: string, public englishName: string, public localizedName: string, public hasLatinCharacters: boolean) {}
 }